+++ src/libshoopdaloop/internal/WithCommandQueue.h
@@ -5,7 +5,7 @@
 // the form of functors can be inserted into the queue and executed
 // in another thread.
 // Typical use is with the audio back-end's process thread.
-template<size_t QueueSize, size_t QueueTimeoutMs, size_t PollIntervalUs>
+template<uint32_t QueueSize, uint32_t QueueTimeoutMs, uint32_t PollIntervalUs>
 class WithCommandQueue {
     CommandQueue ma_queue;
