+++ src/libshoopdaloop/internal/midi_helpers.h
@@ -4,11 +4,11 @@
 #include <optional>
 #include <stdint.h>
 
-size_t channel(const uint8_t *msg_data);
+uint32_t channel(const uint8_t *msg_data);
 
-size_t note(const uint8_t *msg_data);
+uint32_t note(const uint8_t *msg_data);
 
-size_t velocity(const uint8_t *msg_data);
+uint32_t velocity(const uint8_t *msg_data);
 
 bool is_noteOn(const uint8_t *msg_data);
 
@@ -22,6 +22,6 @@ bool is_channel_pressure(const uint8_t *msg_data);
 
 bool is_pitch_wheel(const uint8_t *msg_data);
 
-std::optional<size_t> is_all_notes_off_for_channel(const uint8_t *msg_data);
+std::optional<uint32_t> is_all_notes_off_for_channel(const uint8_t *msg_data);
 
-std::optional<size_t> is_all_sound_off_for_channel(const uint8_t *msg_data);
+std::optional<uint32_t> is_all_sound_off_for_channel(const uint8_t *msg_data);
